+++ src/pytest_docker/__init__.py
@@ -1,5 +1,6 @@
 from .plugin import (
     docker_cleanup,
+    docker_compose_command,
     docker_compose_file,
     docker_compose_project_name,
     docker_ip,